fix: fix ui verify email
