+++ package.json
@@ -3,11 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
     "@hookform/resolvers": "^3.4.0",
-    "@mui/lab": "^5.0.0-alpha.170",
-    "@mui/material": "^5.15.18",
     "@reduxjs/toolkit": "^2.2.5",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
