+++ src/pages/SignUp/signup.module.scss
@@ -28,4 +28,18 @@
             font-weight: 600;
         }
     }
+    .back-login {
+        margin-top: 12px;
+        text-align: right;
+        a {
+            font-size: 15px;
+            text-decoration: underline;
+            cursor: pointer;
+
+            transition: $transition;
+            &:hover {
+                color: $mainColor;
+            }
+        }
+    }
 }
