+++ src/pages/index.ts
@@ -2,7 +2,6 @@ import SignInPage from './SignIn';
 import SignUpPage from './SignUp';
 import ForgotPasswordPage from './ForgotPassword';
 import AddNewSurveyPage from './AddNewSurvey';
-import EmailVerificationPage from './EmailVerification';
 import EmailVerificationResultPage from './EmailVerificationResult';
 import ResetPasswordPage from './ResetPassword';
 import AddNewSurveyQuestionPage from './AddNewSurveyQuestion';
@@ -13,7 +12,6 @@ export {
     SignUpPage,
     ForgotPasswordPage,
     AddNewSurveyPage,
-    EmailVerificationPage,
     EmailVerificationResultPage,
     ResetPasswordPage,
     AddNewSurveyQuestionPage,
